chore(import): add a rate limiter binding for eBird imports
Import is the heaviest write path an account can reach, and it had no limiter
while geocoding did. The key helper moves to its own module rather than being
imported from geocoding-gateway, which is where it happened to live.

Anonymous callers share a budget per IP rather than per account, because
anonymous accounts can be minted without limit and a per-account key would hand
out a fresh allowance each time. Nothing uses the binding yet.
