+++ functions/lib/geocoding-gateway.ts
@@ -10,20 +10,8 @@ import {
 const GEOAPIFY_ORIGIN = 'https://api.geoapify.com'
 const GEOAPIFY_DEADLINE_MS = 5_000
 
-/**
- * Rate limit key for a geocoding request. Registered accounts get their own
- * budget; anonymous sessions share one per IP, because they can be minted
- * without limit and a per-account key would hand out a fresh allowance each time.
- */
-export function rateLimitKey(
-  user: { id?: string; isAnonymous?: boolean } | undefined,
-  request: Request,
-): string {
-  if (user?.isAnonymous) {
-    return `ip:${request.headers.get('cf-connecting-ip') || 'unknown'}`
-  }
-  return `user:${user?.id || 'unknown'}`
-}
+// Re-exported so the geocoding routes keep their existing import.
+export { rateLimitKey } from './rate-limit'
 
 type Fetcher = typeof fetch
 export type GeocodingStage = 'places lookup' | 'reverse fallback' | 'search'
