+++ wrangler.toml
@@ -27,6 +27,17 @@ namespace_id = "1001"
   limit = 20
   period = 60
 
+# Import is the heaviest write path an account can reach. This throttles bursts;
+# the real protection against a forged confirm is signing the preview ids.
+# One import spends two requests here, preview and confirm.
+[[ratelimits]]
+name = "IMPORT_LIMITER"
+namespace_id = "1003"
+
+  [ratelimits.simple]
+  limit = 40
+  period = 60
+
 
 [vars]
 
@@ -65,6 +76,14 @@ namespace_id = "1002"
 
   [env.preview.ratelimits.simple]
   limit = 20
+  period = 60
+
+[[env.preview.ratelimits]]
+name = "IMPORT_LIMITER"
+namespace_id = "1004"
+
+  [env.preview.ratelimits.simple]
+  limit = 40
   period = 60
